Make upgradeslots into model attached to build
